finite tutte le funzionalità dell'admin, aggiunte e sistemate quasi tutte le sessioni

+++ codice/backend/deleteBici.php
@@ -2,17 +2,26 @@
 include_once("../gestioneDB/gestioneDatabase.php");
 
 header('Content-Type: application/json');
+if(!isset($_SESSION))
+{
+    session_start();
+}
+if(isset($_SESSION["isAdmin"]))
+{
+    $gest = new gestioneDatabase();
+    $gest->connettiDb();
+    $id = $_GET['id'];
 
-$gest = new gestioneDatabase();
-$gest->connettiDb();
-$id = $_GET['id'];
+    $risposta = false;
 
-$risposta = false;
+    $risposta = $gest->deleteBici($id);
 
-$risposta = $gest->deleteBici($id);
+    $gest->conn->close();
+    echo json_encode(["status" => $risposta]);
+}
+else
+    echo json_encode(["status" => false, "message" => "Non sei un admin"]);
 
-$gest->conn->close();
-echo json_encode(["status" => $risposta]);
 
 ?>
